booting is now delayed if the RTC is in use
This is intended to fix the issue where the device tries to boot
when the battery is still charging (and still *very* low) and it
gets stuck is a strange state. The most commonly seen symptom is
that the LEDs get stuck on or the device is constantly rebooting.
Neither of these is helpful and they both can keep the battery
from charging further.